Moved threads to separate namespace
