+++ src/stamina/builder/threads/BookKeeperThread.cpp
@@ -1,0 +1,9 @@
+#include "BookKeeperThread.h"
+
+namespace stamina {
+namespace builder {
+namespace threads {
+
+}
+}
+}
